% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{commit}
\alias{commit}
\title{Reexported function from git2r}
\description{
See \code{\link[git2r]{commit}}
}
\seealso{
Other version_control: \code{\link{auto_commit}},
  \code{\link{pull}}, \code{\link{push}},
  \code{\link{recent_commit}}, \code{\link{repository}},
  \code{\link{status}}
}
\concept{version_control}
