% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package.R
\name{data_package}
\alias{data_package}
\title{Create a Data Package for a directory of CSV files}
\usage{
data_package(path = ".")
}
\arguments{
\item{path}{the directory in which to create the \code{datapackage.json} file.}
}
\description{
Create a \code{datapackage.json} file for a directory of CSV files.
The function will look for all \code{.csv} files in the directory and its
subdirectories.
It will then create a \code{datapackage.json} file with the metadata of each CSV
file.
}
\seealso{
Other storage: 
\code{\link{display_metadata}()},
\code{\link{list_data}()},
\code{\link{prune_meta}()},
\code{\link{read_vc}()},
\code{\link{relabel}()},
\code{\link{rename_variable}()},
\code{\link{rm_data}()},
\code{\link{update_metadata}()},
\code{\link{verify_vc}()},
\code{\link{write_vc}()}
}
\concept{storage}
