% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_metadata.R
\name{display_metadata}
\alias{display_metadata}
\title{Display metadata for a \code{git2rdata} object}
\usage{
display_metadata(x, minimal = FALSE)
}
\description{
Display metadata for a \code{git2rdata} object
}
