% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git2rdata-package.R
\docType{package}
\name{git2rdata-package}
\alias{git2rdata}
\alias{git2rdata-package}
\title{git2rdata: Store and retrieve data.frames in a git repository}
\description{
Make versioning of data.frame easy and efficient using git repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/git2rdata}
  \item Report bugs at \url{https://github.com/inbo/git2rdata/issues}
}

}
\author{
\strong{Maintainer}: Thierry Onkelinx \email{thierry.onkelinx@inbo.be} (0000-0001-8804-4216)

Other contributors:
\itemize{
  \item  Research Institute for Nature and Forest \email{info@inbo.be} [copyright holder, funder]
}

}
\keyword{internal}
