% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_changes.R
\name{list_data_changes}
\alias{list_data_changes}
\title{List Commits Where Data Changed}
\usage{
list_data_changes(root = ".", path = NULL, ...)
}
\arguments{
\item{root}{the \code{root} of the repository. Either a path or a \code{git-repository}}

\item{path}{Path to the data file (or another file in the current repo). If
no file extension is provided, \code{.tsv} is added.}

\item{...}{
  Arguments passed on to \code{\link[git2r:commits]{git2r::commits}}
  \describe{
    \item{\code{repo}}{a path to a repository or a \code{git_repository}
object. Default is '.'}
    \item{\code{topological}}{Sort the commits in topological order (parents
before children); can be combined with time sorting. Default
is TRUE.}
    \item{\code{time}}{Sort the commits by commit time; Can be combined with
topological sorting. Default is TRUE.}
    \item{\code{reverse}}{Sort the commits in reverse order; can be combined
with topological and/or time sorting. Default is FALSE.}
    \item{\code{n}}{The upper limit of the number of commits to output. The
default is NULL for unlimited number of commits.}
    \item{\code{ref}}{The name of a reference to list commits from e.g. a tag
or a branch. The default is NULL for the current branch.}
  }}
}
\value{
A data frame with information about the commits that changed the
repository or a file in the repository. Contains columns \code{sha}, \code{message},
\code{summary}, \code{author_name}, \code{author_time}, and others.
}
\description{
Lists commits where the file in \code{path} was changed.
}
\examples{
# Use a temporary directory for this example
dir.create(tmpdir <- tempfile())
owd <- setwd(tmpdir)

# Create our first data commit
set.seed(42)
x <- data.frame(x = 1:5, y = runif(5))

git2r::init()
write_vc(x, "example", sorting = "x")
git2r::add(path = "example*")
git2r::commit(message = "First commit")
git2r::tag(name = "v0.0.1")

cat("Unrelated changes", file = "README.md")
git2r::add(path = "README.md")
git2r::commit(message = "Second commit (non-data)")

x <- rbind(data.frame(x = 6:10, y = runif(5)))
write_vc(x, "example", stage = TRUE)
git2r::add(path = "example*")
git2r::commit(message = "Third commit")

list_data_changes()[c("sha", "message", "author_time")]
list_data_changes(path  = "example")[c("sha", "message", "author_time")]

# clean up
setwd(owd)
unlink(tmpdir, recursive = TRUE)

}
