% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.git2rdata}
\alias{print.git2rdata}
\title{Print method for \code{git2rdata} objects.}
\usage{
\method{print}{git2rdata}(x, ...)
}
\arguments{
\item{x}{a \code{git2rdata} object}

\item{...}{additional arguments passed to \code{print}}
}
\description{
Prints the data and the description of the columns when available.
}
