% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_tsv}
\alias{prune_tsv}
\alias{prune_tsv.git_repository}
\title{Prune tsv files}
\usage{
prune_tsv(root = ".", path = NULL, recursive = TRUE, ...)

\method{prune_tsv}{git_repository}(root, path = NULL, recursive = TRUE,
  ..., stage = FALSE)
}
\arguments{
\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{path}{the directory in which to clean all the data files}

\item{recursive}{remove files in subdirectories too}

\item{...}{additional parameters used in some methods}

\item{stage}{stage the changes after removing the files. Defaults to FALSE.}
}
\value{
returns invisibily a vector of removed files names. The paths are
relative to \code{root}.
}
\description{
Removes all tsv files from the \code{path} when they have an accompanying yml file
}
\seealso{
Other storage: \code{\link{prune_yml}},
  \code{\link{read_vc}}, \code{\link{write_vc}}
}
\concept{storage}
