% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{pull}
\alias{pull}
\title{Reexported function from git2r}
\description{
See \code{\link[git2r]{pull}}
}
\seealso{
Other version_control: \code{\link{commit}},
  \code{\link{push}}, \code{\link{recent_commit}},
  \code{\link{repository}}, \code{\link{status}}
}
\concept{version_control}
