% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{repository}
\alias{repository}
\alias{status}
\alias{commit}
\alias{pull}
\alias{push}
\title{Reexported functions from git2r}
\description{
\itemize{
\item \code{\link[git2r]{repository}}
\item \code{\link[git2r]{pull}}
\item \code{\link[git2r]{status}}
\item \code{\link[git2r]{commit}}
\item \code{\link[git2r]{push}}
}
}
