% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{Relabel factor levels}
\usage{
relabel(file, root, change)
}
\arguments{
\item{file}{the name of the file without file extension. Can include a
relative path. It is relative to the \code{root}.}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{change}{a named list with named vectors. The name of list elements must
match the names of the variables. The names of the vector elements must match
the existing factor labels. The values represent the new factor labels}
}
\value{
invisible \code{NULL}
}
\description{
Imaging the situation where we have a dataframe with a factor variable and we
have stored it with \code{write_vc(optimize = TRUE)}. The raw data file contains
the factor indices and the metadata contains the link between the factor
index and the corresponding label.
}
\examples{
root <- tempfile("git2rdata-relabel")
dir.create(root)
ds <- ds <- data.frame(a = c("a1", "a2"), b = c("b2", "b1"))
write_vc(ds, "relabel", root, sorting = "b")
new_labels <- list(
  a = list(a2 = "a3")
)
relabel("relabel", root, new_labels)
}
