% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{repository}
\alias{repository}
\title{Reexported function from git2r}
\description{
See \code{\link[git2r]{repository}}
}
\seealso{
Other version_control: \code{\link{auto_commit}},
  \code{\link{commit}}, \code{\link{pull}},
  \code{\link{push}}, \code{\link{recent_commit}},
  \code{\link{status}}
}
\concept{version_control}
