% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_data.R
\name{rm_data}
\alias{rm_data}
\alias{rm_data.git_repository}
\title{Remove data files}
\usage{
rm_data(root = ".", path = NULL, type = c("tsv", "yml", "both"),
  recursive = TRUE, ...)

\method{rm_data}{git_repository}(root, path = NULL, type = c("tsv",
  "yml", "both"), recursive = TRUE, ..., stage = FALSE)
}
\arguments{
\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{path}{the directory in which to clean all the data files}

\item{type}{which file type should be removed}

\item{recursive}{remove files in subdirectories too}

\item{...}{additional parameters used in some methods}

\item{stage}{stage the changes after removing the files. Defaults to FALSE.}
}
\value{
returns invisibily a vector of removed files names. The paths are
relative to \code{root}.
}
\description{
Remove all tsv and/or yml files within the path
}
\seealso{
Other storage: \code{\link{read_vc}},
  \code{\link{write_vc}}
}
\concept{storage}
