% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vc.R
\name{root_at_ref}
\alias{root_at_ref}
\title{Locate Project Root at Git Reference}
\usage{
root_at_ref(root = ".", ref = "HEAD")
}
\arguments{
\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (\code{"."}).}

\item{ref}{A \code{git} reference: a branch name, tag name, or commit hash (sha).}
}
\value{
A \code{git_tree} object to be used by \code{read_vc()} to read a \code{git2rdata}
file at a particular point in time.
}
\description{
Returns a \code{git_tree} object representing the project root at the given
reference, typically a branch name or a commit SHA. When combined with
\code{\link[=read_vc]{read_vc()}}, this function lets users load data at a particular point in time
(specifically at a given commit snapshot) without having to explicitly check
out the reference using \code{git} directly.
}
\examples{

# Use a temporary directory for this example
dir.create(tmpdir <- tempfile())
owd <- setwd(tmpdir)

# Create our first data commit
set.seed(42)
x <- data.frame(x = 1:5, y = runif(5))

git2r::init()
write_vc(x, "example", sorting = "x")
git2r::add(path = "example*")
git2r::commit(message = "First commit")
git2r::tag(name = "v0.0.1")

x <- rbind(data.frame(x = 6:10, y = runif(5)))
write_vc(x, "example", stage = TRUE)
git2r::add(path = "example*")
git2r::commit(message = "Second commit")

root_at_ref(ref = "v0.0.1")
read_vc("example", root_at_ref(ref = "v0.0.1"))

}
