% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.git2rdata}
\alias{summary.git2rdata}
\title{Summary method for \code{git2rdata} objects.}
\usage{
\method{summary}{git2rdata}(object, ...)
}
\arguments{
\item{object}{a \code{git2rdata} object}

\item{...}{additional arguments passed to \code{summary}}
}
\description{
Prints the summary of the data and the description of the columns when
available.
}
\seealso{
Other internal: 
\code{\link{is_git2rdata}()},
\code{\link{is_git2rmeta}()},
\code{\link{meta}()},
\code{\link{print.git2rdata}()},
\code{\link{upgrade_data}()}
}
\concept{internal}
