% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_commit.R
\docType{methods}
\name{auto_commit}
\alias{auto_commit}
\alias{auto_commit,ANY,ANY-method}
\alias{auto_commit,ANY,git_repository-method}
\alias{auto_commit,git_repository-methods}
\title{Commit staged changes in a git repository with automated message}
\usage{
auto_commit(package, repo = ".", push = FALSE, ...)

\S4method{auto_commit}{ANY,ANY}(package, repo = ".", push = FALSE, ...)

\S4method{auto_commit}{ANY,git_repository}(package, repo = ".",
  push = FALSE, ...)
}
\arguments{
\item{package}{The name of the package from which we auto commit}

\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{push}{Logical indicating whether to push the commit. Defaults to FALSE.}

\item{...}{arguments passed to \code{\link[git2r]{commit}} and \code{\link[git2r]{push}}}
}
\description{
The mesagge is based on the information returned by \code{\link[utils]{sessionInfo}}
}
