% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_commit.R
\docType{methods}
\name{auto_commit}
\alias{auto_commit}
\title{Commit staged changes in a git repository with automated message}
\usage{
auto_commit(package, repo = ".", push = FALSE, ...)
}
\arguments{
\item{package}{The name of the package from which we auto commit}

\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{push}{Logical indicating whether to push the commit. Defaults to FALSE.}

\item{...}{arguments passed to \code{\link[git2r]{commit}} and
\code{\link[git2r]{push}}}
}
\value{
The \code{git_repository} object that contains the commit
}
\description{
The mesagge is based on the information returned by
\code{\link[utils]{sessionInfo}}
}
\seealso{
Other version_control: \code{\link{commit}},
  \code{\link{pull}}, \code{\link{push}},
  \code{\link{recent_commit}}, \code{\link{repository}},
  \code{\link{status}}
}
\concept{version_control}
