% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git2rdata-package.R
\docType{package}
\name{git2rdata-package}
\alias{git2rdata}
\alias{git2rdata-package}
\title{git2rdata: Store and Retrieve Data.frames in a Git Repository}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Make versioning of data.frame easy and efficient using git repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/git2rdata}
  \item \url{https://doi.org/10.5281/zenodo.1485309}
  \item Report bugs at \url{https://github.com/ropensci/git2rdata/issues}
}

}
\author{
\strong{Maintainer}: Thierry Onkelinx \email{thierry.onkelinx@inbo.be} (0000-0001-8804-4216)

Other contributors:
\itemize{
  \item Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (0000-0002-6378-6229) [contributor]
  \item Peter Desmet \email{peter.desmet@inbo.be} (0000-0002-8442-8025) [contributor]
  \item Research Institute for Nature and Forest \email{info@inbo.be} [copyright holder, funder]
}

}
\keyword{internal}
