% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_git2rmeta.R
\name{is_git2rmeta}
\alias{is_git2rmeta}
\title{Check whether a git2rdata file has valid metadata.}
\usage{
is_git2rmeta(file, root = ".", validate)
}
\arguments{
\item{file}{the name of the file without file extension. Can include a
relative path. \code{file} is a path relative to the \code{root}.}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{validate}{Should invalid metadata result in an error. Defaults to
\code{FALSE}.}
}
\value{
A logical value. \code{TRUE} in case of a valid metadata file. Otherwise
\code{FALSE} or an error depending on \code{validate}
}
\description{
Valid metadata is a file with \code{.yml} extension. It has a top level item
\code{..generic}. This item contains \code{git2rdata} (the version number), \code{hash} (a
hash on the metadata) and \code{data_hash} (a hash on the data file). The version
number must be the current version.
}
\seealso{
Other internal: \code{\link{meta}},
  \code{\link{upgrade_data}}
}
\concept{internal}
