% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{methods}
\name{meta}
\alias{meta}
\title{Optimize a vector for storage as plain text and add meta data}
\usage{
meta(x, optimize = TRUE)
}
\arguments{
\item{x}{the vector}

\item{optimize}{recode the data to get smaller text files. Defaults to TRUE}
}
\value{
the optimized vector \code{x} with \code{meta} attribute
}
\description{
\code{\link{write_vc}} applies this function automatically on your
data.frame.
}
\concept{internal}
