% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,character-method}
\alias{meta,integer-method}
\alias{meta,numeric-method}
\alias{meta,factor-method}
\alias{meta,logical-method}
\alias{meta,complex-method}
\alias{meta,POSIXct-method}
\alias{meta,Date-method}
\title{Optimize a vector for storage as plain text and add meta data}
\usage{
meta(x, optimize = TRUE)

\S4method{meta}{character}(x, optimize = TRUE)

\S4method{meta}{integer}(x, optimize = TRUE)

\S4method{meta}{numeric}(x, optimize = TRUE)

\S4method{meta}{factor}(x, optimize = TRUE)

\S4method{meta}{logical}(x, optimize = TRUE)

\S4method{meta}{complex}(x, optimize = TRUE)

\S4method{meta}{POSIXct}(x, optimize = TRUE)

\S4method{meta}{Date}(x, optimize = TRUE)
}
\arguments{
\item{x}{the vector}

\item{optimize}{recode the data to get smaller text files. Defaults to TRUE}
}
\value{
the optimized vector \code{x} with \code{meta} attribute
}
\description{
Optimize a vector for storage as plain text and add meta data
}
\seealso{
Other storage: \code{\link{read_vc}},
  \code{\link{rm_data}}, \code{\link{write_vc}}
}
\concept{storage}
