% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_meta}
\alias{prune_meta}
\alias{prune_meta.git_repository}
\title{Prune metadata files}
\usage{
prune_meta(root = ".", path = NULL, recursive = TRUE, ...)

\method{prune_meta}{git_repository}(root, path = NULL,
  recursive = TRUE, ..., stage = FALSE)
}
\arguments{
\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{path}{the directory in which to clean all the data files}

\item{recursive}{remove files in subdirectories too}

\item{...}{additional parameters used in some methods}

\item{stage}{stage the changes after removing the files. Defaults to FALSE.}
}
\value{
returns invisibily a vector of removed files names. The paths are
relative to \code{root}.
}
\description{
Removes all metadata (\code{.yml} files) from the \code{path} when they don't have accompanying data (\code{.tsv} file).
}
\examples{
# on file system
root <- tempfile("git2rdata-")
dir.create(root)
write_vc(iris[1:6, ], "iris", root, sorting = "Sepal.Length")
list_data(root)
list.files(root, recursive = TRUE)
rm_data(root, path = ".")
list.files(root, recursive = TRUE)
list_data(root)
prune_meta(root, path = ".")
list.files(root, recursive = TRUE)
list_data(root)

# on git repo
repo_path <- tempfile("git2rdata-repo-")
dir.create(repo_path)
repo <- git2r::init(repo_path)
git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
status(repo)
list_data(repo)
auto_commit(package = "my_package", repo)
status(repo)
rm_data(repo, path = ".")
list_data(repo)
status(repo)
prune_meta(repo, path = ".")
list_data(repo)
status(repo)
}
\seealso{
Other storage: \code{\link{read_vc}},
  \code{\link{relabel}}, \code{\link{rm_data}},
  \code{\link{write_vc}}
}
\concept{storage}
