% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{push}
\alias{push}
\title{Re-exported Function From \code{git2r}}
\description{
See \code{\link[git2r]{push}} in  \code{git2r}.
}
\seealso{
Other version_control: 
\code{\link{commit}()},
\code{\link{pull}()},
\code{\link{recent_commit}()},
\code{\link{repository}()},
\code{\link{status}()}
}
\concept{version_control}
