% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vc.R
\name{read_vc}
\alias{read_vc}
\title{Read a \code{data.frame} from a repository}
\usage{
read_vc(file, root = ".")
}
\arguments{
\item{file}{the name of the file without file extension. Can include a
relative path. It is relative to the \code{root}.}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}
}
\value{
The \code{data.frame}
}
\description{
Read a \code{data.frame} from a repository
}
\seealso{
Other storage: \code{\link{prune_tsv}},
  \code{\link{prune_yml}}, \code{\link{write_vc}}
}
\concept{storage}
