% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vc.R
\name{read_vc}
\alias{read_vc}
\title{Read a \code{data.frame} from a repository}
\usage{
read_vc(file, root = ".")
}
\arguments{
\item{file}{the name of the file without file extension. Can include a
relative path. It is relative to the \code{root}.}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}
}
\value{
The \code{data.frame} with the file names and hashes as attributes
}
\description{
Read a \code{data.frame} from a repository
}
\examples{
# on file system
root <- tempfile("git2rdata-")
dir.create(root)
write_vc(iris[1:6, ], file = "iris", root = root, sorting = "Sepal.Length")
list.files(root, recursive = TRUE)
read_vc("iris", root)
write_vc(iris[1:5, ], "iris", root)
list.files(root, recursive = TRUE)
write_vc(
  iris[1:6, -2], "iris", root, sorting = "Sepal.Length", override = TRUE
)
list.files(root, recursive = TRUE)
write_vc(iris[1:6, ], "iris", root, sorting = "Sepal.Width", override = TRUE)
list.files(root, recursive = TRUE)
write_vc(
  iris[1:6, ], "iris2", root, sorting = "Sepal.Width", optimize = FALSE
)
list.files(root, recursive = TRUE)

# on git repo
repo_path <- tempfile("git2rdata-repo-")
dir.create(repo_path)
repo <- git2r::init(repo_path)
git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
write_vc(iris[1:6, ], file = "iris", root = repo, sorting = "Sepal.Length")
status(repo)
read_vc("iris", repo)
write_vc(iris[1:5, ], "iris", repo, stage = TRUE)
status(repo)
write_vc(
  iris[1:6, -2], "iris", repo, sorting = "Sepal.Length", override = TRUE
)
status(repo)
write_vc(
  iris[1:6, ], "iris", repo, sorting = "Sepal.Width", override = TRUE,
  stage = TRUE
)
status(repo)
write_vc(
  iris[1:6, ], "iris2", repo, sorting = "Sepal.Width", optimize = FALSE
)
status(repo)
}
\seealso{
Other storage: \code{\link{prune_meta}},
  \code{\link{rm_data}}, \code{\link{write_vc}}
}
\concept{storage}
