% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{Relabel factor levels by updating the metadata}
\usage{
relabel(file, root = ".", change)
}
\arguments{
\item{file}{the name of the file without file extension. Can include a
relative path. \code{file} is a path relative to the \code{root}.}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{change}{either a \code{list} or a \code{data.frame}. In case of a \code{list`` is a named }list\code{with named}vectors\code{. The names of list elements must match the names of the variables. The names of the vector elements must match the existing factor labels. The values represent the new factor labels. In case of a }data.frame\code{it needs to have the variables}factor\code{(name of the factor),}old\code{(the old) factor label and}new\code{(the new factor label).}relabel()` ignores all other columns.}
}
\value{
invisible \code{NULL}
}
\description{
Imagine the situation where we have a dataframe with a factor variable and we
have stored it with \code{write_vc(optimize = TRUE)}. The raw data file contains
the factor indices and the metadata contains the link between the factor
index and the corresponding label. In such a case, relabeling a factor can be
fast and lightweight by updating the metadata.
}
\examples{

# setup a directory
root <- tempfile("git2rdata-relabel")
dir.create(root)

# create a dataframe and store it
ds <- ds <- data.frame(a = c("a1", "a2"), b = c("b2", "b1"))
write_vc(ds, "relabel", root, sorting = "b")

# define new labels as a list and apply them
new_labels <- list(
  a = list(a2 = "a3")
)
relabel("relabel", root, new_labels)

# define new labels as a dataframe and apply them
change <- data.frame(
  factor = c("a", "a", "b"),
  old = c("a3", "a1", "b2"),
  new = c("c2", "c1", "b3")
)
relabel("relabel", root, change)

# clean up
junk <- file.remove(list.files(root, full.names = TRUE), root)
}
\seealso{
Other storage: \code{\link{list_data}},
  \code{\link{prune_meta}}, \code{\link{read_vc}},
  \code{\link{rm_data}}, \code{\link{write_vc}}
}
\concept{storage}
