% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{rm_data}
\alias{rm_data}
\alias{rm_data.git_repository}
\title{Remove data files}
\usage{
rm_data(root = ".", path = NULL, recursive = TRUE, ...)

\method{rm_data}{git_repository}(root, path = NULL, recursive = TRUE,
  ..., stage = FALSE, type = c("unmodified", "modified", "ignored",
  "all"))
}
\arguments{
\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (".").}

\item{path}{the directory in which to clean all the data files}

\item{recursive}{remove files in subdirectories too}

\item{...}{additional parameters used in some methods}

\item{stage}{stage the changes after removing the files. Defaults to FALSE.}

\item{type}{which classes of files should be removed. \code{unmodified} are files in the git history and unchanged since the last commit. \code{modified} are files in the git history and changed since the last commit. \code{ignored} refers to file listed in a \code{.gitignore} file. Selecting \code{modified} will remove both \code{unmodified} and \code{modified} data files. Selecting \code{ìgnored} will remove \code{unmodified}, \code{modified} and \code{ignored} data files. \code{all} refers to all visible data files, inclusing \code{untracked} files. The argument can be abbreviated to the first letter.}
}
\value{
returns invisibily a vector of removed files names. The paths are
relative to \code{root}.
}
\description{
Removes all data (\code{.tsv} files) from the \code{path} when they have accompanying metadata (\code{.yml} file). The metadata remains untouched.
}
\seealso{
Other storage: \code{\link{prune_meta}},
  \code{\link{read_vc}}, \code{\link{write_vc}}
}
\concept{storage}
