% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_data.R
\docType{methods}
\name{rm_data}
\alias{rm_data}
\alias{rm_data,character-method}
\alias{rm_data,git_repository-method}
\title{Remove data files
Remove all tsv and/or yml files within the path}
\usage{
rm_data(root, path = NULL, type = c("tsv", "yml", "both"),
  recursive = TRUE, ...)

\S4method{rm_data}{character}(root, path = NULL, type = c("tsv", "yml",
  "both"), recursive = TRUE, ...)

\S4method{rm_data}{git_repository}(root, path = NULL, type = c("tsv",
  "yml", "both"), recursive = TRUE, ..., stage = FALSE)
}
\arguments{
\item{root}{The root of a project. Can be a file path or a \code{git-repository}}

\item{path}{the directory in which to clean all the data files}

\item{type}{which file type should be removed}

\item{recursive}{remove files in subdirectories too}

\item{...}{additional parameters used in some methods}

\item{stage}{stage the changes after writing the data. Defaults to FALSE}
}
\description{
Remove data files
Remove all tsv and/or yml files within the path
}
