% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{status}
\alias{status}
\title{Re-exported Function From \code{git2r}}
\description{
See \code{\link[git2r]{status}} in \code{git2r}.
}
\seealso{
Other version_control: 
\code{\link{commit}()},
\code{\link{pull}()},
\code{\link{push}()},
\code{\link{recent_commit}()},
\code{\link{repository}()}
}
\concept{version_control}
