% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vc.R
\docType{methods}
\name{write_vc}
\alias{write_vc}
\alias{write_vc,ANY,ANY,character-method}
\title{Write a \code{data.frame} to a git repository
This will create two files. The \code{".tsv"} file contains the raw data.
The \code{".yml"} contains the meta data on the columns in YAML format.}
\usage{
write_vc(x, file, root, sorting, override = FALSE, optimize = TRUE, ...)

\S4method{write_vc}{ANY,ANY,character}(x, file, root, sorting,
  override = FALSE, optimize = TRUE, ...)
}
\arguments{
\item{x}{the \code{data.frame}}

\item{file}{the name of the file without file extension. Can include a relative
path. It is relative to the \code{root}.}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}}

\item{sorting}{a vector of column names defining which columns to use for
sorting \code{x} and in what order to use them. Defaults to
\code{colnames(x)}}

\item{override}{Ignore existing meta data. This is required when new
variables are added or variables are deleted. Setting this to TRUE can
potentially lead to large diffs. Defaults to FALSE.}

\item{optimize}{recode the data to get smaller text files. Defaults to TRUE}

\item{...}{additional parameters used in some methods}
}
\value{
a named vector with the hashes of the files. The names contains the
files with their paths relative to the root of the git_repository.
}
\description{
Write a \code{data.frame} to a git repository
This will create two files. The \code{".tsv"} file contains the raw data.
The \code{".yml"} contains the meta data on the columns in YAML format.
}
